move servers instanciations in their corresponding backend package
Instead of having these awkward declaration side-by-side, this allows for a better division of what's needed by
one package or the other, and better show what actually is the common code between both.
