+++ lib/core/cardano-wallet-core.cabal
@@ -41,7 +41,6 @@ library
     , cardano-addresses
     , cardano-crypto
     , cborg
-    , conduit
     , containers
     , contra-tracer
     , cryptonite
@@ -59,7 +58,6 @@ library
     , generic-lens
     , http-api-data
     , http-client
-    , http-conduit
     , http-media
     , http-types
     , iohk-monitoring
@@ -100,7 +98,6 @@ library
     , x509
     , x509-store
     , x509-validation
-    , zip
     , Win32-network
   hs-source-dirs:
       src
@@ -108,6 +105,7 @@ library
       Cardano.Byron.Codec.Cbor
       Cardano.DB.Sqlite
       Cardano.DB.Sqlite.Delete
+      Cardano.Pool
       Cardano.Pool.DB
       Cardano.Pool.DB.MVar
       Cardano.Pool.DB.Model
