+++ lib/jormungandr/test/unit/Cardano/Pool/Jormungandr/MetadataSpec.hs
@@ -12,14 +12,14 @@ import Cardano.Pool.Jormungandr.Metadata
     , MetadataConfig (..)
     , RegistryLog (..)
     , RegistryLogMsg (..)
-    , StakePoolMetadata (..)
-    , StakePoolTicker
     , cacheArchive
     , getMetadataConfig
     , getStakePoolMetadata
     )
+import Cardano.Wallet.Api.Types
+    ( ApiT (..) )
 import Cardano.Wallet.Primitive.Types
-    ( PoolOwner (..) )
+    ( PoolOwner (..), StakePoolMetadata (..), StakePoolTicker )
 import Cardano.Wallet.Unsafe
     ( unsafeFromText )
 import Codec.Archive.Zip
@@ -301,7 +301,7 @@ stakePoolEntryMeta (Meta m) = Just m
 
 encodeStakePoolEntry :: StakePoolEntry -> ByteString
 encodeStakePoolEntry Junk = "junk"
-encodeStakePoolEntry (Meta m) = BL.toStrict $ encode m
+encodeStakePoolEntry (Meta m) = BL.toStrict $ encode (ApiT m)
 
 isJunk :: StakePoolEntry -> Bool
 isJunk Junk = True
